# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marine Corps logistics calculator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List all units
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{unitId}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Get unit detail
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unit detail with supply summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitDetail"
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update a unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete a unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Echelon:
      type: string
      enum: [fireteam, squad, section, platoon, company, battalion, regiment, battery, team]
    UnitRole:
      type: string
      enum: [organic, attached, in_support]
      description: Relationship of the unit to the regiment (organic subordinate, attached, or in direct support)
    AmmoPosture:
      type: string
      enum: [combat_load, assault, sustain]
      description: Doctrinal ammo planning posture. combat_load = initial issue quantity; assault = daily assault rate; sustain = daily sustain rate.
...
    UnitWeapon:
      type: object
      properties:
        id:
          type: string
        unitId:
          type: string
        weaponSystemId:
          type: string
        weaponName:
          type: string
        tamcn:
          type: string
          nullable: true
        quantity:
          type: integer
          minimum: 1
        isGce:
          type: boolean
      required: [id, unitId, weaponSystemId, weaponName, quantity, isGce]
...
    Climate:
      type: string
      enum: [temperate, arid, tropical, arctic]
    OpTempo:
      type: string
      enum: [garrison, sustained, high, combat]
    SupplyClass:
      type: string
      description: Marine Corps Class of Supply
      enum: ["I", "III", "V", "VIII", "IX"]
    Unit:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        callsign:
          type: string
          nullable: true
        echelon:
          $ref: "#/components/schemas/Echelon"
        personnel:
          type: integer
          minimum: 1
        commander:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        climate:
          $ref: "#/components/schemas/Climate"
        opTempo:
          $ref: "#/components/schemas/OpTempo"
        missionDays:
          type: integer
          minimum: 1
          description: Planned mission duration in days
        readiness:
          type: integer
          minimum: 0
          maximum: 100
          description: Computed readiness percentage (0-100)
        deficiencyCount:
          type: integer
          minimum: 0
        role:
          $ref: "#/components/schemas/UnitRole"
        ammoPosture:
          $ref: "#/components/schemas/AmmoPosture"
        isGce:
          type: boolean
          description: True if this unit is a GCE (Ground Combat Element); affects which doctrinal ammo rates apply
        distroEmails:
          type: array
          items:
            type: string
            format: email
            pattern: '{{DISTRO_EMAIL_PATTERN}}'
          description: Distribution list of email addresses pre-filled when the unit's schedules are emailed (S-4, supporting battalion, logistics POCs, etc.)
        distroCcEmails:
          type: array
          items:
            type: string
            format: email
            pattern: '{{DISTRO_EMAIL_PATTERN}}'
          description: Optional CC distribution list pre-filled into the cc= field of the schedule email — used for visibility recipients (supporting battalion logistics, company commander, higher HQ) who are not action addressees.
        distroBccEmails:
          type: array
          items:
            type: string
            format: email
            pattern: '{{DISTRO_EMAIL_PATTERN}}'
          description: Optional BCC distribution list pre-filled into the bcc= field of the schedule email — used for silent recipients (operations watch officer, archive mailbox, COC liaison) who should not be visible to To or CC recipients.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - echelon
        - personnel
        - climate
        - opTempo
        - missionDays
        - readiness
        - deficiencyCount
        - role
        - ammoPosture
        - isGce
        - distroEmails
        - distroCcEmails
        - distroBccEmails
        - createdAt
        - updatedAt
    UnitInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        callsign:
          type: string
          nullable: true
        echelon:
          $ref: "#/components/schemas/Echelon"
        personnel:
          type: integer
          minimum: 1
        commander:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        climate:
          $ref: "#/components/schemas/Climate"
        opTempo:
          $ref: "#/components/schemas/OpTempo"
        missionDays:
          type: integer
          minimum: 1
        role:
          $ref: "#/components/schemas/UnitRole"
        ammoPosture:
          $ref: "#/components/schemas/AmmoPosture"
        isGce:
          type: boolean
        distroEmails:
          type: array
          items:
            type: string
            format: email
            pattern: '{{DISTRO_EMAIL_PATTERN}}'
          description: Distribution list of email addresses pre-filled when the unit's schedules are emailed
        distroCcEmails:
          type: array
          items:
            type: string
            format: email
            pattern: '{{DISTRO_EMAIL_PATTERN}}'
          description: Optional CC distribution list pre-filled into the cc= field of the schedule email
        distroBccEmails:
          type: array
          items:
            type: string
            format: email
            pattern: '{{DISTRO_EMAIL_PATTERN}}'
          description: Optional BCC distribution list pre-filled into the bcc= field of the schedule email — silent recipients (operations watch officer, archive mailbox, COC liaison) hidden from To and CC recipients
      required: [name, echelon, personnel, climate, opTempo, missionDays]
    CatalogItem:
      type: object
      properties:
        id:
          type: string
        supplyClass:
          $ref: "#/components/schemas/SupplyClass"
        name:
          type: string
        nsn:
          type: string
          nullable: true
          description: National Stock Number
        unit:
          type: string
          description: Unit of issue (e.g. "MRE", "gal", "battery", "round", "L")
        baseDailyRate:
          type: number
          description: Base consumption per Marine per day at temperate / sustained tempo
        criticality:
          type: string
          enum: [low, medium, high, critical]
        notes:
          type: string
          nullable: true
        isCustom:
          type: boolean
          description: True if this item was added manually by a planner (not from the standard catalog)
        createdAt:
          type: string
          format: date-time
          description: When this catalog item was first created
        updatedAt:
          type: string
          format: date-time
          description: When this catalog item's details were last edited (equal to createdAt if never edited)
        usedByUnitCount:
          type: integer
          minimum: 0
          description: Number of units that currently track this catalog item on their supply list. Populated by listCatalogItems; may be omitted on other endpoints.
        usedByUnits:
          type: array
          description: Units that currently track this catalog item, sorted by unit name. Populated by listCatalogItems; may be omitted on other endpoints.
          items:
            $ref: "#/components/schemas/CatalogItemUsageUnit"
      required:
        - id
        - supplyClass
        - name
        - unit
        - baseDailyRate
        - criticality
        - isCustom
        - createdAt
        - updatedAt
    CatalogItemUsageUnit:
      type: object
      description: Minimal unit reference for catalog item usage lists
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]
...
    SupplyEntry:
      type: object
      properties:
        id:
          type: string
        unitId:
          type: string
        itemId:
          type: string
        item:
          $ref: "#/components/schemas/CatalogItem"
        onHand:
          type: number
          minimum: 0
        dailyConsumption:
          type: number
          description: Adjusted daily consumption for this unit (climate / tempo / personnel)
        daysOfSupply:
          type: number
          description: Computed days of supply remaining
        required:
          type: number
          description: Effective required quantity (requiredOverride when set, otherwise auto-computed)
        shortfall:
          type: number
          description: Required minus on-hand (>= 0)
        status:
          type: string
          enum: [green, amber, red]
        requiredOverride:
          type: number
          nullable: true
          description: Planner-set required quantity for this item on this unit. Null means use the auto-computed value. Zero means not a requirement.
        isRequirement:
          type: boolean
          description: False when requiredOverride is exactly 0 — item is tracked but not counted toward readiness or DOS.
        burnBreakdown:
          type: string
          nullable: true
          description: Human-readable description of how the burn rate was computed (weapon-driven for Class V)
        combatLoadTarget:
          type: number
          nullable: true
          description: Total combat load target quantity (weapon-driven Class V only)
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - unitId
        - itemId
        - item
        - onHand
        - dailyConsumption
        - daysOfSupply
        - required
        - shortfall
        - status
        - isRequirement
        - updatedAt
...
    UnitDetail:
      type: object
      properties:
        unit:
          $ref: "#/components/schemas/Unit"
        supplyByClass:
          type: array
          items:
            $ref: "#/components/schemas/ClassSummary"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/SupplyEntry"
        upcomingResupply:
          type: array
          items:
            $ref: "#/components/schemas/ResupplyEvent"
        weapons:
          type: array
          items:
            $ref: "#/components/schemas/UnitWeapon"
      required: [unit, supplyByClass, entries, upcomingResupply, weapons]
    ClassSummary:
      type: object
      properties:
        supplyClass:
          $ref: "#/components/schemas/SupplyClass"
        label:
          type: string
        itemCount:
          type: integer
        worstDaysOfSupply:
          type: number
        status:
          type: string
          enum: [green, amber, red]
      required: [supplyClass, label, itemCount, worstDaysOfSupply, status]
...
    ResupplyEvent:
      type: object
      properties:
        id:
          type: string
        unitId:
          type: string
        unitName:
          type: string
        supplyClass:
          $ref: "#/components/schemas/SupplyClass"
        itemId:
          type: string
          nullable: true
        itemName:
          type: string
          nullable: true
        quantity:
          type: number
        unit:
          type: string
        scheduledFor:
          type: string
          format: date-time
        status:
          type: string
          enum: [projected, planned, in_transit, delivered, cancelled]
        assignedTo:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - unitId
        - unitName
        - supplyClass
        - quantity
        - unit
        - scheduledFor
        - status
        - createdAt
...